requirements.txt
opencv-python
numpy 
onnxruntime 
scipy 
tqdm 
mediapipe
paho-mqtt